Make earnings smarter (#6)
This commit changes the way earnings are calculated. Initially we calculated
based on the current net hashrate. Now, earnings utilizes the difficulty over
the last eight hours for a somewhat more stable calculation.

I wanted to do 24 hours but couldn't quickly figure out how to get that out of
the insight api.